feat(verifier): support compressed proofs besides checking vkey (#848)
* uncomment sp1-verifier

* copy compressed verifier from turbo, commented out

* get recursion machine to compile on wasm besides errno (from sp1-jit)

* remove vestigial sp1_core_path attr from MachineAir derive

* move poseidon2 operation from sp1-core-machine to sp1-hypercube

* move some util functions from sp1-core-machine to sp1-hypercube

* remove dependency on sp1-core-machine from sp1-recursion-machine

* merge imports and format

* remove dependency on sp1-core-machine from sp1-recursion-executor

* minimize and clean up all tokio feature flags

* uncomment alloc/core imports in compressed verifier

* make `MemoryPermitting` work for all pointer widths

* document poseidon2 operation

* move `SP1RecursionProof` to `sp1-hypercube`

* compressed verifier, everything except vk verification

* clean up commented out vkey verification stuff

* update `sp1_verifier_valid` test

* split off internal module in compressed verifier

* fix lints for utility functions moved to `sp1-hypercube`

* use `tokio/full` in `sp1-cli`

* clippy fixes

+++ crates/core/machine/src/executor.rs
@@ -37,7 +37,7 @@ pub struct MachineExecutor<F: PrimeField32> {
 }
 
 impl<F: PrimeField32> MachineExecutor<F> {
-    pub fn new(record_buffer_size: u64, num_record_workers: usize, opts: SP1CoreOpts) -> Self {
+    pub fn new(record_buffer_size: usize, num_record_workers: usize, opts: SP1CoreOpts) -> Self {
         let machine = RiscvAir::<F>::machine();
 
         Self {
